contract

+++ src/main/java/com/cg/fms/repository/UserRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cg.fms.entities.User;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
 	User findByUserName(String userName);
 
@@ -17,4 +17,7 @@ public interface UserRepository extends JpaRepository<User, Long> {
 	User findByEmail(String email);
 	
 	Optional<User> findByEmailAndPassword(String email , String password);
+	
+	
+	
 }
